+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.2.0 (in development)
 -----------------------
 - Require the `port` field of `SMTPSender` to be non-negative
 - Mark `Sender` as `runtime_checkable` and export it
+- Gave the `outgoing` command `--section` and `--no-section` options
 
 v0.1.0 (2021-03-06)
 -------------------