+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 v0.2.0 (in development)
 -----------------------
 - Require the ``port`` field of ``SMTPSender`` to be non-negative
+- Mark `Sender` as ``runtime_checkable`` and export it
+- Gave the :command:`outgoing` command ``--section`` and ``--no-section``
+  options
 
 v0.1.0 (2021-03-06)
 -------------------